requirements.txt
PyQt6==6.9.1
PyQt6-Qt6==6.9.2
PyQt6_sip==13.10.2
websockets>=12.0
# Authentication & database
sqlalchemy>=2.0
psycopg2-binary>=2.9
alembic>=1.13
python-dotenv>=1.0
